requirements.txt
fastapi
pyautogen
python-dotenv
sqlmodel
uvicorn